added loading screens, dark mode and ects as double

+++ lib/domain/general_domain/entities/general_course_entity.dart
@@ -2,7 +2,7 @@ class GeneralCourseEntity {
   final String uniId;
   final String programId;
   final String name;
-  final int ects;
+  final double ects;
   final String field;
   final String id;
   final bool graded;
